refactor: move verifyGuild
