+++ apps/bot/src/lib/epic-helper/command-helper/guild/index.ts
@@ -1,7 +1,9 @@
 import {_renderMultipleGuildEmbed} from './embed/multiple-guild';
 import {_renderThisWeekUpgraidListEmbed} from './embed/this-week-upgraid-list';
+import {_verifyGuild} from './verify-guild';
 
 export const _guildHelper = {
   renderMultipleGuildEmbed: _renderMultipleGuildEmbed,
   renderThisWeekUpgraidList: _renderThisWeekUpgraidListEmbed,
+  verifyGuild: _verifyGuild,
 };
