+++ apps/bot/src/lib/epic-rpg/commands/guild/_shared.ts
@@ -1,12 +1,9 @@
-import type {Client, Embed, Guild, User} from 'discord.js';
+import type {Client, Embed, User} from 'discord.js';
 import {EmbedBuilder} from 'discord.js';
 import {upgraidService} from '../../../../services/database/upgraid.service';
 import {guildService} from '../../../../services/database/guild.service';
 import {_renderThisWeekUpgraidListEmbed} from '../../../epic-helper/command-helper/guild/embed/this-week-upgraid-list';
 import {djsMessageHelper} from '../../../discordjs/message';
-import commandHelper from '../../../epic-helper/command-helper';
-import type {IGuild} from '@epic-helper/models';
-import {djsMemberHelper} from '../../../discordjs/member';
 
 interface ISendRecordsToGuildChannel {
   guildServerId: string;
@@ -57,77 +54,4 @@ export const _sendUpgraidResultToGuildChannel = async ({
     },
     client,
   });
-};
-
-interface IVerifyGuild {
-  client: Client;
-  server: Guild;
-  userId: string;
-}
-
-export const verifyGuild = async ({userId, client, server}: IVerifyGuild) => {
-  const roles = await getUserGuildRoles({
-    client,
-    userId,
-    server,
-  });
-  const userGuild = await guildService.findUserGuild({
-    userId,
-  });
-  let finalGuild: IGuild | null = userGuild;
-  let embed: EmbedBuilder | null = null;
-
-  if (roles && roles.size > 1) {
-    embed = commandHelper.guild.renderMultipleGuildEmbed(roles);
-  } else if (roles?.size === 1) {
-    const guildRole = roles.first()!;
-    const guild = await guildService.findGuild({
-      serverId: server.id,
-      roleId: guildRole.id,
-    });
-
-    if (userGuild?.roleId !== guildRole.id)
-      await guildService.registerUserToGuild({
-        userId,
-        serverId: server.id,
-        roleId: guildRole.id,
-      });
-    if (guild) finalGuild = guild;
-  }
-
-  if (userGuild && userGuild.serverId === server.id && !roles?.size) {
-    await guildService.removeUserFromGuild({
-      serverId: server.id,
-      roleId: userGuild.roleId,
-      userId,
-    });
-  }
-
-  return {
-    guild: finalGuild,
-    errorEmbed: embed,
-  };
-};
-
-export interface IGetUserGuildRoles {
-  client: Client;
-  server: Guild;
-  userId: string;
-}
-
-export const getUserGuildRoles = async ({
-  server,
-  userId,
-  client,
-}: IGetUserGuildRoles) => {
-  const serverMember = await djsMemberHelper.getMember({
-    serverId: server.id,
-    client,
-    userId,
-  });
-  if (!serverMember) return null;
-  const guildRoles = await guildService.getAllGuildRoles({serverId: server.id});
-  return serverMember.roles.cache.filter((userRole) =>
-    guildRoles.some((guildRole) => userRole.id === guildRole),
-  );
 };
